Change view_manager/public/interfaces to use relative paths.
This CL introduces the pattern for making services' public interfaces flexible
with respect to their location in a client repo: the BUILD.gn file and mojoms
use relative rather than absolute paths to reference their dependencies. Note
that we assume that the relative structure of a service and its dependencies is
fixed; it is only the location of all of those services *as a group* that can
change in a client repo.

To bootstrap services' BUILD.gn files being able to refer to files under the
Mojo SDK, this CL adds the requirement that the parent directory of services'
client code contain a mojo_sdk_root.gni file that defines a mojo_sdk_root
variable pointing to the parent directory of the Mojo SDK.

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/798923002