Dart: Adds validation conformance testing.
This change runs the conformance test suite for message format
validation. It fixes a few bugs for the tests to pass.

The error messages produced are not yet conformant. For now I
just check that some exception is thrown from the Decoder on an
ill-formed message.

BUG=
R=abarth@chromium.org

Review URL: https://codereview.chromium.org/888763004