Python: Move common code out of the example application
This CL moves ApplicationImpl and ServiceProviderImpl out of the example Python
app into the public SDK. It defines a new ApplicationDelegate interface, which
is what apps will use to customize the behavior of ApplicationImpl. The code in
MojoMain is moved into a RunMojoApplication() function, similar to
ApplicationRunner in C++.

This CL is in preparation for introducing a second Python application into the
codebase.

R=qsr@chromium.org

Review URL: https://codereview.chromium.org/928433004