Adding @promise.async annotation.
This allow to transform a function f such that all its argument will be resolved if they are promises, and it will transform its returns value into a promise.

This allows to easily chain calls with promises.

For example if you define:

@promise.async
def add(x, y):
  x + y

You can then call add with either x or y as promise, and it will return a promise that will compute the sum.

R=blundell@chromium.org

Review URL: https://codereview.chromium.org/964873002