Fix non-compiling generated code for array union-member validation.
Validation for the active member of a union is performed inside a switch
statement which was made of a single block. But we need to initialize a
variable containing the array validation parameters for that array.
So a block needs to be created within which the initialization is performed.
(Otherwise, the C++ standard won't let us jump to a later label.)

BUG=
R=yzshen@chromium.org

Review URL: https://codereview.chromium.org/1060063005