Java Jinja templates: Check for method response params being None instead of empty.
The significant change in this CL is that in a boolean check I replaced:

  "method.response_parameters"

with

  "method.response_parameters != None".

This resolves the following problem: In the case that there are zero arguments in the response, the Java Jinja templates were generating code that returns a response without the MESSAGE_IS_RESPONSE flag.

Additionally I renamed a couple of Jinja macro variables.

R=abarth@chromium.org, qsr@chromium.org, qsr
BUG=#329

Review URL: https://codereview.chromium.org/1242193002 .