Simplify transformation from data to module in mojom code generation pipeline.
The mojom code generation pipeline includes a step where a module.Module object is created. This object contains a list of imports. The imports used to contain a reference to the module.Module corresponding to the import. In this CL I instead create a minimalist copy of the module.Module in question containing only the needed data. The purpose is to make the data dependencies explicit.

R=rudominer@chromium.org

Review URL: https://codereview.chromium.org/1370733003 .