Move mojo::internal::MessageBuilder out of the |internal| namespace.
This can be used to build Message (e.g interface messages) from the bindings.
It might be just as easy to build Messages by constructing a |Message| manually and store it in a byte buffer, but MessageBuilder helps construct flags in a safer way (e.g., Messages can't be a response *and* expect a response).

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1460463003 .