Add a dummy mojo_system to allow the use of the python bindings for serialization only.
If any attribute of the dummy module is accessed, it will throw an error informing the user of the need to build mojo_system.

The new mojom bindings generator will use this in order to run without having to first build mojo_system. This will make running on Mac easier.

BUG= #461
R=rudominer@chromium.org

Review URL: https://codereview.chromium.org/1468143003 .