Simplify Go's Mojom Type Generator
Removes all dependence on the identifier_store.

Originally, the identifier_store appeared to be needed because types were generated recursively, and their children types also needed to be registered. In order to break cycles and avoid repeats, we would use the identifier_store.

However, this is actually not needed. Each relevant type to be registered (enum/union/struct/interface) is already known beforehand by the generating Python script. Types belonging to other .mojom files are also registered explicitly. Since children types do NOT need to be registered, the identifier_store could be safely removed.

This CL also shifts the generation flag from the go generator to the generator scripts. This is a better location since it allows users to control whether types are generated or not.

BUG=https://github.com/domokit/mojo/issues/596
R=azani@chromium.org

Review URL: https://codereview.chromium.org/1508293004 .