Mojom compiler: Uniformizes the name lookup algorithm.
Previously our name lookup algorithm had two undesirable properties:
- Names for values and names for types were registered and looked up separately.
- Names for non-referents (e.g. fields of a struct) were not registered in a scope at all.

The result was that we could not detect a name collision between, for example, a type and a struct field. There are cases where this is bad. See the examples in https://github.com/domokit/mojo/issues/620.

In this patch we uniformize name registration and lookup and we extend name registration to other object types (such as struct fields) even though those object types are not referents in Mojom.

- In scopes.go we replace the two maps |typesByname| and |valuesByName| with a single map |declaredObjects|.

- The value type for |declaredObjects| is the interface |DeclaredObject| and in addition to user-defined types
and user-defined-values there are now other objects that implement this interfaces such as StructField and Method.

- Our lookup algorithm has been made more general within scopes.go: The method LookupObject() takes a parameter
|acceptFilter| which is a bitmask of the types of objects that should be accepted during a lookup. This gives
us a flexible infrastructure that will allow us to tweak the precise lookup algorithm in the future.

- In the next layer up, in mojom_descriptor.go, we implement the new lookup algorithm: We pass in the
bitmask |LookupAcceptAll|. This is how we get the new uniform lookup behavior. Because the lower layer is flexible we can easily tweak this in the future if we decide to.

- There was some special logic to handle the case of looking up a name that might turn out to be an enum value.
This logic has been moved up from scopes.go into the higher layer in mojom_descriptor.go

BUG=#620,#504
R=azani@chromium.org

Review URL: https://codereview.chromium.org/1590403006 .