Mojo C++ bindings: Generate InterfaceHandle<> instead of InterfacePtr<>.
This includes replacing all consumers of InterfacePtr<> (which were
previously generated; mostly they are implementations of interfaces)
with InterfaceHandle<>.

This was a fairly mechanical process of replacing InterfacePtr with
InterfaceHandle<>, and I've transformed one from the other at the
earliest occurance.  There are places where I could've delayed
converting InterfaceHandle->InterfacePtr to a later point, but
this entire job was very arduous.

This CL needs to follow up with another doing the same thing, but in
Flutter engine's services.

BUG=#662
R=viettrungluu@chromium.org, jamesr@chromium.org

Review URL: https://codereview.chromium.org/1682113003 .